게시판 제목 불러오는 함수를 utility로 추출
