+++ src/components/pages/board/BoardHeader.tsx
@@ -1,8 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { ChevronDown } from 'lucide-react';
-import { firestore } from '../../../firebase';
-import { doc, getDoc } from 'firebase/firestore';
+import { fetchBoardTitle } from '../../../utils/boardUtils';
 
 interface BoardHeaderProps {
   boardId: string | undefined;
@@ -12,30 +11,12 @@ const BoardHeader: React.FC<BoardHeaderProps> = ({ boardId }) => {
   const [title, setTitle] = useState<string>('Loading...');
 
   useEffect(() => {
-    const fetchBoardTitle = async () => {
-      try {
-        if (!boardId) {
-          console.error('No boardId provided');
-          setTitle('Board not found');
-          return;
-        }
-        
-        const boardDocRef = doc(firestore, 'boards', boardId);
-        const boardDoc = await getDoc(boardDocRef);
-        if (boardDoc.exists()) {
-          const boardData = boardDoc.data();
-          setTitle(boardData?.title || 'Board');
-        } else {
-          console.error('Board not found');
-          setTitle('Board not found');
-        }
-      } catch (error) {
-        console.error('Error fetching board title:', error);
-        setTitle('Error loading title');
-      }
+    const loadBoardTitle = async () => {
+      const title = await fetchBoardTitle(boardId || '');
+      setTitle(title);
     };
 
-    fetchBoardTitle();
+    loadBoardTitle();
   }, [boardId]);
 
   return (
